Move atoms to folders
