+++ src/pages/NewYear/atoms/NewYearCategoryAtom/NewYearCategoryAtom.jsx
