+++ src/pages/NewYear/atoms/NewYearDateAtom/NewYearDateAtom.jsx
@@ -1,0 +1,14 @@
+import { ParagraphAtom } from '../../../../components/atoms/ParagraphAtom/ParagraphAtom.jsx';
+
+import css from './NewYearDateAtom.module.css';
+
+function NewYearDateAtom() {
+  return (
+    <>
+      <ParagraphAtom className={css.newYear__date} text={'31/12/2024'} />
+      <ParagraphAtom className={css.newYear__date} text={'20:00'} />
+    </>
+  );
+}
+
+export { NewYearDateAtom };
